Convert portage ansible module to use portage python module
The ansible portage module uses the gentoolkit utility equery to perform
lookups for packages. gentoolkit is not included in the @system set and
may not be available to all systems.

The portage package manager, as it is written in python, has a python module
that can perform the queries that ansible requires, and is available on
all systems that have portage installed.